+++ packages/app/src/components-styled/cms/rich-content.tsx
@@ -2,7 +2,7 @@ import { PortableTextEntry } from '@sanity/block-content-to-react';
 import { ReactNode } from 'react';
 import { Box } from '~/components-styled/base';
 import { ContentBlock } from '~/components-styled/cms/content-block';
-import { ContentImage } from '~/components-styled/cms/image';
+import { ContentImage } from '~/components-styled/cms/content-image';
 import { getFileAssetSrc, PortableText } from '~/lib/sanity';
 import { InlineAttachment } from '~/types/cms';
 import { assert } from '~/utils/assert';
