+++ packages/app/src/components-styled/editorial-detail.tsx
@@ -2,7 +2,7 @@ import css from '@styled-system/css';
 import ArrowIcon from '~/assets/arrow.svg';
 import { Box } from '~/components-styled/base';
 import { ContentBlock } from '~/components-styled/cms/content-block';
-import { ContentImage } from '~/components-styled/cms/image';
+import { ContentImage } from '~/components-styled/cms/content-image';
 import { Heading } from '~/components-styled/typography';
 import { PortableText } from '~/lib/sanity';
 import siteText from '~/locale';
